Changed cur_ambient and cur_diffuse from a colorRGBA to a colorRGB since the alpha component is unused.

+++ src/Tree.cpp
@@ -1125,7 +1125,7 @@ void tree::update_leaf_orients() { // leaves move in wind or when struck by an o
 			}
 			wscale = dot_product(local_wind, leaves[i].norm);
 		}
-		if (i < leaf_cobjs.size()) { // rotate leaves when hit by an object (plus another i update hack)
+		if (i < (int)leaf_cobjs.size()) { // rotate leaves when hit by an object (plus another i update hack)
 			coll_obj &cobj(get_leaf_cobj(i));
 				
 			if (cobj.last_coll > 0) {
